requirements.txt
argon2-cffi==20.1.0
asgiref==3.3.1
attrs==20.3.0
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
click==7.1.2
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.4.6
defusedxml==0.7.1
Django==3.1.7
django-cors-headers==3.7.0
django-debug-toolbar==3.2
django-extensions==3.1.1
django-filter==2.4.0
django-model-utils==4.1.1
django-money==1.3.1
django-simpleui==2021.4
django-templated-mail==1.1.1
djangorestframework==3.12.2
djangorestframework-simplejwt==4.6.0
djoser==2.1.0
drf-yasg==1.20.0
gunicorn==20.0.4
h11==0.12.0
idna==2.10
inflection==0.5.1
iniconfig==1.1.1
itypes==1.2.0
Jinja2==2.11.3
Markdown==3.3.4
MarkupSafe==1.1.1
oauthlib==3.1.0
packaging==20.9
pluggy==0.13.1
py==1.10.0
py-moneyed==0.8.0
pycparser==2.20
PyJWT==2.0.1
pyparsing==2.4.7
pytest==6.2.2
pytest-django==4.1.0
python3-openid==3.2.0
pytz==2021.1
requests==2.25.1
requests-oauthlib==1.3.0
ruamel.yaml==0.16.13
ruamel.yaml.clib==0.2.2
six==1.15.0
social-auth-app-django==4.0.0
social-auth-core==4.1.0
sqlparse==0.4.1
toml==0.10.2
uritemplate==3.0.1
urllib3==1.26.4
uvicorn==0.13.4
whitenoise==5.2.0
